chore: restructure code - dashboard views - payments
